- modifications due to changed aspect definition in AOP XML file - implemented cascading of enable/disable functionality (IN PROGRESS)
git-svn-id: https://svn.code.sf.net/p/openuss/code/branches/solarplexus@2213 60046477-0c10-0410-bf4b-e552a4ff1768

+++ openuss-plexus/plexus/plexus-core/src/main/java/org/openuss/lecture/CourseIndexingAspect.java
@@ -32,7 +32,7 @@ public class CourseIndexingAspect {
 	public void createCourseIndex(CourseInfo courseInfo) {
 		logger.info("Starting method createCourseIndex");
 		try {
-			if (courseInfo.getAccessType() != AccessType.CLOSED) {
+			if (courseInfo.isEnabled() && courseInfo.getAccessType() != AccessType.CLOSED) {
 				course = courseDao.courseInfoToEntity(courseInfo);
 				indexerService.createIndex(course);
 			}
@@ -45,13 +45,13 @@ public class CourseIndexingAspect {
 	public void updateCourseIndex(CourseInfo courseInfo) {
 		logger.info("Starting method updateCourseIndex");
 		try {
-			if (courseInfo.getAccessType() == AccessType.CLOSED) {
+			if (courseInfo.isEnabled() && courseInfo.getAccessType() != AccessType.CLOSED) {
+				course = courseDao.courseInfoToEntity(courseInfo);
+				indexerService.updateIndex(course);
+			} else {
 				logger.info("Deleting CourseIndex");
 				course = courseDao.courseInfoToEntity(courseInfo);
 				indexerService.deleteIndex(course);
-			} else {
-				course = courseDao.courseInfoToEntity(courseInfo);
-				indexerService.updateIndex(course);
 			}
 		} catch (IndexerApplicationException e) {
 			logger.error(e);
